+++ geniuspace/app/Models/NodeI18n.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class NodeI18n extends Model
+{
+    protected $table = 'node_i18n';
+    public $incrementing = false;
+    public $timestamps = false;
+    protected $guarded = [];
+}