+++ geniuspace/app/Models/Thread.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Thread extends Model
+{
+    public $incrementing = false;
+    public $timestamps = false;
+    protected $keyType = 'string';
+    protected $guarded = [];
+}