+++ geniuspace/app/Models/WikiPage.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class WikiPage extends Model
+{
+    public $timestamps = false;
+    protected $guarded = [];
+}